Initial work on Drumfish to boot an ATmega128RFA1
Initial work on a project I'm calling Drumfish where the goal is to boot
a complete ATmega128RFA1 board.  Currently the code is laid out with the
goal to allow other boards in the future. The UART misbehaves currently
as it does not buffer and flushes a byte at a time which makes it not
behave like real chips in the wild.