Refactor UART to remove unused bits and misc fixes
Went through the UART code and removed functionality that I'll never
want to expose (a tap with a CRLF conversion). Fixed a few bugs like the
assumption that fd=0 is not a valid fd. If we daemonize and open a UART
then the fd we'll get is 0. Changed code style to match the rest of the
project. Added a lot of error checking and reporting in those cases.