uart: disable arbitary usleep(1) in simavr
SimAVR runs the UARTs with a usleep(1) by default but we definitely
don't want that behavior since it messes up the UART timing. The example
firmware expects to send a byte every 200ms but with the usleep(1) its
closer to 5s which is a huge time delay. Without the usleep(1) the
timing as fast as the system will go.